src/main/java/br/com/carvalho/CarvalhoAtelieAdm/service/produto/ValidarProdutoService.java
package br.com.carvalho.CarvalhoAtelieAdm.service.produto;

import br.com.carvalho.CarvalhoAtelieAdm.domain.Produto;
import br.com.carvalho.CarvalhoAtelieAdm.exception.ProdutoNaoEncontradoException;
import br.com.carvalho.CarvalhoAtelieAdm.repository.ProdutoRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class ValidarProdutoService {

    @Autowired
    private ProdutoRepository produtoRepository;

    public Produto validarEObterProduto(Long id) {
        return produtoRepository.findById(id)
                .orElseThrow(() -> new ProdutoNaoEncontradoException("Produto não encontrado com o ID: " + id));
    }
}